moved auth type to interface and renamed, updated all related imports
